Enable bugprone-narrowing-conversions
510 sites at the tighter setting, cleared with no suppressions -- not the 839
this document predicted, because the sign-compare and truncation commits had
already taken 329 of them.

The check is enabled with WarnOnEquivalentBitWidth: false, recorded in
.clang-tidy with the reason. At its defaults it also reports 585 same-width
unsigned-to-signed conversions, which are LEPTON (unsigned short), COORDINATE
(uint32_t) and the rest of the do-not-touch list in docs/TYPE_MIGRATION.md
flowing into int. Those are findings the project has already decided against;
configuring the check is honest, suppressing 585 sites would not be.

Worked in passes from safest shape to riskiest, building and auditing after
each one:

  single-line declarations, CELL x = EXPR;               510 -> 438
  assignments and compound assignments, typed by the
  diagnostic's target type                              438 -> 210
  the same, allowing trailing comments and *p / arr[i]   210 -> 171
  right-hand sides that already contained a cast         171 -> 127
  column wrap behind a verification gate                 127 ->  21
  the remainder by hand                                   21 ->   0

Two audits ran against the committed files after every batch: each compound
rewrite must expand to exactly 'x = cast(x OP (E))', and every other rewrite
must reduce to its committed line once the inserted cast is stripped. The final
run reports 0 non-equivalent rewrites.

The audit earned its place immediately. The first compound pass wrote

  length -= strparse - strstart - 1;
  length = static_cast<int>(length - strparse - strstart - 1);

which drops the grouping and changes the result. 44 of the 93 compound rewrites
needed the parentheses, and nothing but the comparison against HEAD would have
found them -- the code compiles either way.

The column wrap from cf44023c was reused, but only behind a gate that rejects
any candidate not reducing exactly back to its original: 83 wraps applied, 36
rejected. One still slipped through as 'sizeof(Random) static_cast<int>(-
index)', textually reducible and syntactically invalid, so an operator-start
guard was added after the compiler caught it.

Measures 0 tree-wide afterwards, together with clang-diagnostic-comma and
readability-redundant-casting: no comma-operator damage and no cast left
redundant. Both games build and the 84 tests pass. The full-config sweep was
not run for this commit.

+++ src/ra/abstract_type.cc
@@ -22,7 +22,7 @@ int AbstractTypeClass::Full_Name() const {
   // string in NameOverride rather than the normal text table.
   for (base::ssize index = 0; index < std::ssize(NameOverride); index++) {
     if (NameIDOverride[index] == (RTTI + 1) * 100 + ID) {
-      return -(index + 1);
+      return static_cast<int>(-(index + 1));
     }
   }
   return FullName;
