+++ src/td/cdata.cc
@@ -1309,7 +1309,7 @@ const short* TemplateTypeClass::Occupy_List(bool) const {
   ptr = &_occupy[0];
   for (index = 0; index < Width * Height; index++) {
     if (map[index] != 0xFF) {
-      *ptr++ = index % Width + index / Width * MAP_CELL_W;
+      *ptr++ = static_cast<short>(index % Width + index / Width * MAP_CELL_W);
     }
   }
   *ptr = REFRESH_EOL;
