+++ src/td/sdata.cc
@@ -261,7 +261,7 @@ const short* SmudgeTypeClass::Occupy_List(bool) const {
 
   for (int x = 0; x < Width; x++) {
     for (int y = 0; y < Height; y++) {
-      *ptr++ = x + y * MAP_CELL_W;
+      *ptr++ = static_cast<short>(x + y * MAP_CELL_W);
     }
   }
   *ptr = REFRESH_EOL;
